requirements.txt
# Core dependencies
pydantic>=2.0.0,<3.0.0
fastapi>=0.100.0,<1.0.0
uvicorn[standard]>=0.20.0,<1.0.0
websockets>=11.0.0,<12.0.0

# Authentication and security
PyJWT>=2.8.0,<3.0.0
python-multipart>=0.0.6,<1.0.0

# AWS and AgentCore
boto3>=1.28.0,<2.0.0
botocore>=1.31.0,<2.0.0

# HTTP and web scraping
httpx>=0.24.0,<1.0.0
aiohttp>=3.8.0,<4.0.0
beautifulsoup4>=4.12.0,<5.0.0
lxml>=4.9.0,<5.0.0

# Data processing and analysis
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0,<2.0.0
spacy>=3.6.0,<4.0.0

# NLP and text processing
nltk>=3.8.0,<4.0.0
textblob>=0.17.0,<1.0.0

# Async and concurrency
asyncio-throttle>=1.0.0,<2.0.0
aiofiles>=23.0.0,<24.0.0

# Logging and monitoring
structlog>=23.0.0,<24.0.0
prometheus-client>=0.17.0,<1.0.0

# Testing
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0
pytest-mock>=3.11.0,<4.0.0
httpx-mock>=0.10.0,<1.0.0

# Development tools
black>=23.0.0,<24.0.0
isort>=5.12.0,<6.0.0
flake8>=6.0.0,<7.0.0
mypy>=1.5.0,<2.0.0

# Documentation
mkdocs>=1.5.0,<2.0.0
mkdocs-material>=9.0.0,<10.0.0